add productController authorization

+++ backend/controllers/productsController.js
@@ -2,11 +2,19 @@ import { Router } from "express";
 import { deleteProduct, getProduct, getProducts, newProduct, updateProduct, updateProductStock } from "../services/productsService.js";
 import { getCategoryByName } from "../services/categoriesService.js";
 import { handleError } from "../utils/handleErrors.js";
+import authLoggedUser from "../middlewares/userAuth.js";
 
 const router = Router();
 
-router.post('/', async (req, res) => {
+router.post('/', authLoggedUser, async (req, res) => {
   try {
+    const payload = res.locals.user;
+    if (!payload.isAdmin || !payload.isEmployee) {
+      let error = Error('Only employees can add products');
+      error.status = 405;
+      error.validator = 'Authorization';
+      return handleError(res, error);
+    }
     let category = req.body.category_id;
     category = await getCategoryByName(category);
     req.body = { ...req.body, category_id: category._id };
@@ -19,6 +27,7 @@ router.post('/', async (req, res) => {
 
 router.get('/', async (req, res) => {
   try {
+
     let products = await getProducts();
     res.send(products);
   } catch (error) {
@@ -36,8 +45,15 @@ router.get('/:id', async (req, res) => {
   }
 })
 
-router.put('/:id', async (req, res) => {
+router.put('/:id', authLoggedUser, async (req, res) => {
   try {
+    const payload = res.locals.user;
+    if (!payload.isAdmin || !payload.isEmployee) {
+      let error = Error('Only employees can edit products');
+      error.status = 405;
+      error.validator = 'Authorization';
+      return handleError(res, error);
+    }
     const { id } = req.params;
     let product = await updateProduct(id, req.body);
     res.send(product);
@@ -46,8 +62,15 @@ router.put('/:id', async (req, res) => {
   }
 })
 
-router.patch('/:id', async (req, res) => {
+router.patch('/:id', authLoggedUser, async (req, res) => {
   try {
+    const payload = res.locals.user;
+    if (!payload.isAdmin || !payload.isEmployee) {
+      let error = Error('Only employees can update product stock');
+      error.status = 405;
+      error.validator = 'Authorization';
+      return handleError(res, error);
+    }
     const { id } = req.params;
     let product = await updateProductStock(id, req.body);
     res.send(product);
@@ -56,8 +79,15 @@ router.patch('/:id', async (req, res) => {
   }
 })
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', authLoggedUser, async (req, res) => {
   try {
+    const payload = res.locals.user;
+    if (!payload.isAdmin || !payload.isEmployee) {
+      let error = Error('Only employees can delete products');
+      error.status = 405;
+      error.validator = 'Authorization';
+      return handleError(res, error);
+    }
     const { id } = req.params;
     let product = await deleteProduct(id);
     res.send(product);
